migration double publishing

+++ database/migrations/2023_07_01_463540_create_saves_table.php
@@ -1,6 +1,5 @@
 <?php
 
-
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -9,12 +8,11 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('sales', function (Blueprint $table) {
+        if (Schema::hasTable('saves')) return;
+        Schema::create('saves', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->foreignuuid('user_id');
             $table->foreignuuid('book_id');
-            $table->Integer('quantity');
-            $table->string('status');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
@@ -24,6 +22,6 @@ return new class extends Migration
 
     public function down()
     {
-        Schema::dropIfExists('sales');
+        Schema::dropIfExists('saves');
     }
 };
